SONAR-5329 - Added LogBackendMediumTest and LogIndex updates

+++ sonar-core/src/main/java/org/sonar/core/log/Loggable.java
@@ -28,6 +28,6 @@ public interface Loggable {
 
   Map<String, String> getDetails();
 
-  Long getExecutionTime();
+  Integer getExecutionTime();
 
 }
