+++ sonar-server/src/main/java/org/sonar/server/qualityprofile/ActiveRuleChange.java
@@ -110,7 +110,7 @@ public class ActiveRuleChange implements Loggable {
   }
 
   @Override
-  public Long getExecutionTime() {
+  public Integer getExecutionTime() {
     return null;
   }
 }
