+++ sonar-server/src/test/java/org/sonar/server/log/db/TestLog.java
@@ -44,7 +44,7 @@ public class TestLog implements Loggable {
   }
 
   @Override
-  public Long getExecutionTime() {
+  public Integer getExecutionTime() {
     return null;
   }
 }
